Add Tests and All The Things
Setup a test and dev config, with test employing the use of the newly
imported Mox dep. Added the dictionary file for a list of all
acceptable words and updated the BoardWord module to calculate and
return the largest (x, y) coords for the board size.

A Dictionary module was added to map the dictionary file, reject invalid
cases, and upcase all words. All words that have the given words are returned.
Also a behaviour was implemented that causes modules to either read a file and
return a stream, or take input and return it as a stream. This helps testing so
a prod/dev file that may change won't break tests.