Added Many Code Quality Tools
Addressed an initial error Credo returned concerning an unknown type,
but would still finish successfully - the `Behaviour` for
`DictionaryFileProvider` was referencing `Stream.t()` in its callback,
but what was needed is `File.Stream.t()`

Individually inspected checks in the Credo config file, enabling many
to keep restrictions tighter. In turn, this increased Credo warnings
from 0 to around 50. I commented out the `Readability.Specs` for now,
but intend to add specs and try to stick with that check enabled.

`PipeChainStart` and `SinglePipe` checks were uncommented, many issues
uncovered, and all of them were fixed. Some order rearrangement with
alias', attributes, and similar required adjustments were made. A
moduledoc is now required on all modules as well, even it's just
`@moduledoc false`. Ideally though, this will lead to increased quality
of docs from the top down.

A `.dialyzer_ignore.exs` was created, but currently only contains an
empty list. Language requirements were installed with `asdf` local,
resulting in a `.tool-versions` being committed, the required Credo
version was updated, and dialyzer was further configured.

There is now a `hex_version_check.exs` that checks your current Hex
version and raises with an informative error if it's incorrect.
`git_hooks` are also setup with specific mix tasks running when
triggered by certain git actions, which fail if any of the mix tasks
fail.

Most of the remaining changes were simple reorders of `import`, `alias`
, or similar except the README. The build status was already present,
but now the basic rules and concept of the game have been fleshed out,
detailed explanation around code quality and what's being used to
ensure it stays aligned throughout the codebase. There plenty of file
references paired with an urging for folks to look things over so
"magically aliased" modules don't come as a surprise.

A good bit was also added to address testing, the tools being used,
and the goal of maximum coverage was mentioned. The `git_hooks` that
are in place were explained and files to reference were provided. On
GitHub, the ability to simply merge has been disabled, and preference
is being given to "squash and commit" but with stress on retaining
even the small, seemingly insignificant commits when doing so. This is
in the interest of git history and the potential of needing to
leverage a tool like `git bisect` somewhere down the road.

Lastly, even admins are set to have somewhat limited ability in GH. All
force pushes to `master` have been disallowed, but admins do still have
the ability to squash & merge before checks finish.